requirements.txt
Flask==1.1.1
gunicorn==20.0.2
httplib2==0.9.2
Jinja2==2.10.3
numpy==1.17.4
Pillow==5.1.0
pytesseract==0.3.0
regex==2019.11.1

